# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multiview Control API",
    "description": "REST + WebSocket + SSE management API for the Multiview engine.",
    "contact": {
      "name": "The Multiview Authors"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/alarms": {
      "get": {
        "tags": [
          "alarms"
        ],
        "summary": "`GET /api/v1/alarms` — list active/historical alarms (role: read).",
        "operationId": "list_alarms",
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "description": "Keep only alarms at or above this X.733 severity (e.g. `major`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Keep only active (`true`) or only cleared/historical (`false`) alarms.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "Keep only alarms whose scope kind matches (e.g. `tile`, `probe`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching alarms, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlarmRecordDoc"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alarms/{id}/ack": {
      "post": {
        "tags": [
          "alarms"
        ],
        "summary": "`POST /api/v1/alarms/{id}/ack` — acknowledge an alarm (role: write; If-Match).",
        "operationId": "ack_alarm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Alarm id to acknowledge.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The acknowledged alarm.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmRecordDoc"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to acknowledge.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No alarm with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "description": "If-Match precondition failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "428": {
            "description": "If-Match precondition required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "`GET /api/v1/audit` — list audit entries newest-first (role: read).",
        "description": "An optional `?object_id=` filters to a single object's history.",
        "operationId": "list_audit",
        "parameters": [
          {
            "name": "object_id",
            "in": "query",
            "description": "Restrict the listing to a single object id.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entries, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/layouts": {
      "get": {
        "tags": [
          "layouts"
        ],
        "summary": "`GET /api/v1/layouts` — list all layouts (role: read).",
        "operationId": "list_layouts",
        "responses": {
          "200": {
            "description": "All layouts, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Layout"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outputs": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "`GET /api/v1/outputs` — list all outputs (role: read).",
        "operationId": "list_outputs",
        "responses": {
          "200": {
            "description": "All outputs, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/overlays": {
      "get": {
        "tags": [
          "overlays"
        ],
        "summary": "`GET /api/v1/overlays` — list all overlays (role: read).",
        "operationId": "list_overlays",
        "responses": {
          "200": {
            "description": "All overlays, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salvos": {
      "get": {
        "tags": [
          "salvos"
        ],
        "summary": "`GET /api/v1/salvos` — list salvo definitions (role: read).",
        "operationId": "list_salvos",
        "responses": {
          "200": {
            "description": "All salvo definitions, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalvoDoc"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salvos/{id}": {
      "put": {
        "tags": [
          "salvos"
        ],
        "summary": "`PUT /api/v1/salvos/{id}` — create-or-replace a salvo definition.",
        "description": "On create returns `201`; on replace enforces `If-Match` (→ `412`) and returns\n`200`. The submitted body is validated by [`Salvo::validate`] first; the path\n`id` is authoritative (the body's `id` is overwritten to match it).",
        "operationId": "put_salvo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Salvo id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvoDoc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The replaced salvo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalvoDoc"
                }
              }
            }
          },
          "201": {
            "description": "The created salvo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalvoDoc"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "description": "If-Match precondition failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The salvo failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salvos/{id}/arm": {
      "post": {
        "tags": [
          "salvos"
        ],
        "summary": "`POST /api/v1/salvos/{id}/arm` — arm (stage) the salvo (role: write; 202).",
        "operationId": "arm_salvo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Salvo id to arm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "head",
            "in": "query",
            "description": "The output head this recall targets, if scoped to one head.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Arm accepted; outcome on the realtime stream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to arm.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No salvo with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Engine command bus at capacity; shed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salvos/{id}/cancel": {
      "post": {
        "tags": [
          "salvos"
        ],
        "summary": "`POST /api/v1/salvos/{id}/cancel` — cancel an armed salvo (role: write; 202).",
        "operationId": "cancel_salvo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Salvo id to cancel.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "head",
            "in": "query",
            "description": "The output head this recall targets, if scoped to one head.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Cancel accepted; outcome on the realtime stream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to cancel.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No salvo with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Engine command bus at capacity; shed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "`GET /api/v1/sources` — list all sources (role: read).",
        "operationId": "list_sources",
        "responses": {
          "200": {
            "description": "All sources, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlarmRecordDoc": {
        "type": "object",
        "description": "`OpenAPI` mirror of [`multiview_core::alarm::AlarmRecord`].\n\nSerde-equivalent field-for-field. `id` is the alarm id string;\n`raised_at`/`dwell` are media-time nanoseconds.",
        "required": [
          "id",
          "kind",
          "severity",
          "scope",
          "raised_at",
          "dwell",
          "latched",
          "ack"
        ],
        "properties": {
          "ack": {
            "$ref": "#/components/schemas/AckStateDoc",
            "description": "Operator acknowledgement state."
          },
          "dwell": {
            "type": "integer",
            "format": "int64",
            "description": "How long the condition has persisted, media-time nanoseconds."
          },
          "id": {
            "type": "string",
            "description": "Stable identity of this alarm instance."
          },
          "kind": {
            "$ref": "#/components/schemas/AlarmKindDoc",
            "description": "The probe/fault class that raised it."
          },
          "latched": {
            "type": "boolean",
            "description": "Whether the alarm is latched."
          },
          "raised_at": {
            "type": "integer",
            "format": "int64",
            "description": "When the alarm was first raised, media-time nanoseconds."
          },
          "scope": {
            "$ref": "#/components/schemas/AlarmScopeDoc",
            "description": "What the alarm applies to."
          },
          "severity": {
            "$ref": "#/components/schemas/PerceivedSeverityDoc",
            "description": "Current perceived severity (X.733)."
          }
        }
      },
...
    }
  },
...
}